[ft #[162234065]] Create utilities function for processing all apis requests
