+++ UI/login.html
@@ -47,5 +47,5 @@
       </p>
     </div-->
     <script src="static/js/authentication/main.js"></script>
-    <script src="static/js/authentication/login.js"></script>
+    <script type='module' src="static/js/authentication/login.mjs"></script>
 </html>
